on second thought move StableType back to private agian
