+++ fundamentals/fundamentals-reflection/src/main/scala/izumi/fundamentals/reflection/UniverseGeneric.scala
@@ -3,7 +3,5 @@ package izumi.fundamentals.reflection
 import scala.reflect.api.Universe
 
 trait UniverseGeneric {
-
   val u: Universe
-
 }
